[Timers] Namespacing for non_doc timers for different lambdas
* Presently non_doc timer blobs in metadata bucket were ISO 8601
  timestamps in UTC. This will obviously have problems where there
  are more than one lambdas trying to create non_doc timers. Hence
  namespacing was necessary across non_doc timers for different
  lambdas.

* Leveraging cas for doc_id based timers for xattr update subdoc
  operations. Else there will be missing timers.

Reviewed-on: http://review.couchbase.org/80363
Tested-by: Abhishek Singh <abhishek@couchbase.com>

+++ consumer/control_routine.go
@@ -68,7 +68,7 @@ func (c *Consumer) controlRoutine() {
 			go c.processNonDocTimerEvents()
 
 			if val, ok := settings["deadline_timeout"]; ok {
-				c.socketTimeout = time.Duration(val.(int)) * time.Second
+				c.socketTimeout = time.Duration(val.(float64)) * time.Second
 			}
 
 		case <-c.restartVbDcpStreamTicker.C:
